refactor: rename aiplatform-components to vertex-components for clarity
