+++ components/vertex-components/src/vertex_components/lookup_model.py
