+++ components/vertex-components/src/vertex_components/model_batch_predict.py
